file system api, simple textured meshes support
Now the file system api sits in fs:: namespace. On engine init we scan every application folder and file, load full names for every entry and use SeFileHandle and SeFolderHandle to refer to a file or a folder. This eliminates additional string allocations and simplifies DataProvider logic

Added color textures for mesh assets